telegram-backend-bot/requirements.txt
python-telegram-bot[rate-limiter]==21.6
httpx==0.27.2
websockets==12.0
python-dotenv==1.0.1
